requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
better-profanity==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.0
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
facebook-sdk==3.1.0
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
inflection==0.5.1
Markdown==3.4.4
msgpack==1.0.7
multidict==6.0.4
mysqlclient==2.2.0
packaging==23.2
paypalrestsdk==1.13.3
Pillow==10.0.1
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
twilio==8.9.1
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.0.6
user-agents==2.2.0
uvicorn==0.24.0.post1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
